feat: LevelDB store (#95)
## Relevant issue(s)

Resolves #94

## Description

This PR adds support for a LevelDB backed store.

LevelDB has some limitations around transactions:

> Only one transaction can be opened at a time. Subsequent call to Write
and OpenTransaction will be blocked until in-flight transaction is
committed or discarded. The returned transaction handle is safe for
concurrent use.
> 
> Transaction is expensive and can overwhelm compaction, especially if
transaction size is small. Use with caution.
> 
> The transaction must be closed once done, either by committing or
discarding the transaction. Closing the DB will discard open
transaction.

Because of this we do not create implicit transactions for the typical
`Get`, `Set`, `Delete` and `NewIterator` functions. This should not be a
problem in the browser since JavaScript uses a single threaded event
loop.

This also means the `txn-multi` test multiplier is skipped for level
because it will always timeout due to a transaction deadlock.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/validate-conventional-style.sh](tools/configs/validate-conventional-style.sh).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Added to integration tests.

Specify the platform(s) on which this was tested:
- MacOS