feat: Support transactions via context (#101)
## Relevant issue(s)

Resolves #100

## Description

This PR enables all stores to utilize transactions set via context.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/validate-conventional-style.sh](tools/configs/validate-conventional-style.sh).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Added `txn-context` multiplier to integration tests.

Specify the platform(s) on which this was tested:
- MacOS