fix: LevelDB transaction nil panic (#99)
## Description

I forgot to add a nil check here. This PR also bumps the `goleveldb`
version since it includes a fix for a data race.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/validate-conventional-style.sh](tools/configs/validate-conventional-style.sh).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Integration tests

Specify the platform(s) on which this was tested:
- MacOS