fix: Bump goleveldb version (#104)
## Description

This PR bumps the version of goleveldb. The new version of goleveldb
includes fixes for a data race and deadlock that were discovered from
testing with the DefraDB integration tests. You can view those changes
here:
https://github.com/sourcenetwork/goleveldb/commit/27249d06b81bde61b33a06e481bdb4619ff1b579


## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/validate-conventional-style.sh](tools/configs/validate-conventional-style.sh).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Integration tests.

Specify the platform(s) on which this was tested:
- MacOS